+++ integration-tests/jest/1-create-interactive-analysis.integration-test.js
@@ -1,5 +1,5 @@
 const { registerTest } = require('./jest-utils')
-const { testCreateInteractiveAnalysis } = require('../tests/create-interactive-analysis')
+const { testCreateInteractiveAnalysis } = require('../tests/interactive-analysis/create-interactive-analysis')
 
 
 registerTest(testCreateInteractiveAnalysis)
