+++ integration-tests/jest/1-run-notebook.integration-test.js
@@ -1,5 +1,5 @@
 const { registerTest } = require('./jest-utils')
-const { testRunNotebook } = require('../tests/run-notebook')
+const { testRunNotebook } = require('../tests/interactive-analysis/run-notebook')
 
 
 registerTest(testRunNotebook)
