+++ integration-tests/jest/1-run-workflow-on-snapshot.integration-test.js
@@ -1,5 +1,5 @@
 const { registerTest } = require('./jest-utils')
-const { testRunWorkflowOnSnapshot } = require('../tests/run-workflow-on-snapshot')
+const { testRunWorkflowOnSnapshot } = require('../tests/workspaces/run-workflow-on-snapshot')
 
 
 registerTest(testRunWorkflowOnSnapshot)
