+++ integration-tests/jest/find-workflow.integration-test.js
@@ -1,5 +1,5 @@
 const { registerTest } = require('./jest-utils')
-const { testFindWorkflow } = require('../tests/find-workflow')
+const { testFindWorkflow } = require('../tests/batch-workflows/find-workflow')
 
 
 registerTest(testFindWorkflow)
