+++ integration-tests/jest/import-cohort-data.integration-test.js
@@ -1,5 +1,5 @@
 const { registerTest } = require('./jest-utils')
-const { testImportCohortData } = require('../tests/import-cohort-data')
+const { testImportCohortData } = require('../tests/workspaces/import-cohort-data')
 
 
 registerTest(testImportCohortData)
