+++ integration-tests/jest/import-dockstore-workflow.integration-test.js
@@ -1,5 +1,5 @@
 const { registerTest } = require('./jest-utils')
-const { testImportDockstoreWorkflow } = require('../tests/import-dockstore-workflow')
+const { testImportDockstoreWorkflow } = require('../tests/batch-workflows/import-dockstore-workflow')
 
 
 registerTest(testImportDockstoreWorkflow)
