+++ integration-tests/jest/janitor.integration-test.js
@@ -1,5 +1,5 @@
 const { registerTest } = require('./jest-utils')
-const { janitor } = require('../tests/janitor')
+const { janitor } = require('../tests/workspaces/janitor')
 
 
 registerTest(janitor)
