+++ integration-tests/jest/link-to-new-workspace.integration-test.js
@@ -1,5 +1,5 @@
 const { registerTest } = require('./jest-utils')
-const { testLinkToNewWorkspace } = require('../tests/link-to-new-workspace')
+const { testLinkToNewWorkspace } = require('../tests/data-catalog/link-to-new-workspace')
 
 
 registerTest(testLinkToNewWorkspace)
