+++ integration-tests/jest/preview-dataset.integration-test.js
@@ -1,5 +1,5 @@
 const { registerTest } = require('./jest-utils')
-const { testPreviewDataset } = require('../tests/preview-dataset')
+const { testPreviewDataset } = require('../tests/data-catalog/preview-dataset')
 
 
 registerTest(testPreviewDataset)
