+++ integration-tests/jest/register-user.integration-test.js
@@ -1,5 +1,5 @@
 const { registerTest } = require('./jest-utils')
-const { testRegisterUser } = require('../tests/register-user')
+const { testRegisterUser } = require('../tests/workspaces/register-user')
 
 
 registerTest(testRegisterUser)
