+++ integration-tests/jest/run-catalog-workflow.integration-test.js
@@ -1,5 +1,5 @@
 const { registerTest } = require('./jest-utils')
-const { testCatalog } = require('../tests/run-catalog-workflow')
+const { testCatalog } = require('../tests/data-catalog/run-catalog-workflow')
 
 
 registerTest(testCatalog)
