+++ integration-tests/tests/workspaces/billing-projects.js
@@ -1,7 +1,7 @@
 // This test is owned by the Workspaces Team.
 const _ = require('lodash/fp')
-const { click, clickable, dismissNotifications, findText, noSpinnersAfter, select, signIntoTerra } = require('../utils/integration-utils')
-const { withUserToken } = require('../utils/terra-sa-utils')
+const { click, clickable, dismissNotifications, findText, noSpinnersAfter, select, signIntoTerra } = require('integration-tests/utils/integration-utils')
+const { withUserToken } = require('integration-tests/utils/terra-sa-utils')
 
 
 const billingProjectsPage = (testPage, testUrl) => {
